makes changes to how rows toggle
Changes made to createToggleRow function changes behavior of how rows function. Previously, rows were made up of checkboxes representing stitches, when any stitch was checked, its status would toggle. The changes made in this commit make it so that when a stitch in a row is clicked, all stitches prior to it are marked as complete.

This change allows users to work multiple stitches and then click a single checkbox, no longer requiring them to click for every single stitch. It also accounts for users potentially undoing stitches, so if a user were to undo multiple stitches, they could click on the stitch they returned to, and not have to manually untoggle each stitch.