fixes TempRow Display
tempRow Display can now handle tempInputs composed of multiple types,

for example, it is now possible to add to temp (inc x1, sc x3) x 3 to the temp row.

Previously this would return as the last type in the row and the total amount of stitches, in this example it would show "sc * 12" instead of the aforementioned (inc x1, sc x3) x3. This should make it easier for users to define their patterns and start tracking faster